CLDR-17287 API performance test for the back end
- split out DataDrivenSTTestHandler
